requirements.txt
azure-functions==1.20.0
azure-eventhub==5.11.5
